refactor(missions): refactor the Controller to include the ResponseEntity in my routes and customize the responses to client requests

+++ src/main/java/dev/java10x/ninjasregister/Missions/MissionsController.java
@@ -1,5 +1,8 @@
 package dev.java10x.ninjasregister.Missions;
+import dev.java10x.ninjasregister.Ninjas.NinjaDTO;
 import dev.java10x.ninjasregister.Ninjas.NinjaModel;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -16,32 +19,52 @@ public class MissionsController {
 
     // Show all missions (READ)
     @GetMapping("/list")
-    public List<MissionsDTO> listMissions() {
-        return missionsService.listMissions();
+    public ResponseEntity<List<MissionsDTO>> listMissions() {
+        List<MissionsDTO> missions = missionsService.listMissions();
+        return ResponseEntity.ok(missions);
     }
 
     // Show mission by rank (READ)
     @GetMapping("/list/{rank}")
-    public List<MissionsDTO> getMissionByRank(@PathVariable String rank) {
-        return missionsService.getMissionByRank(rank);
+    public ResponseEntity<?> getMissionByRank(@PathVariable String rank) {
+        List<MissionsDTO> missions = missionsService.getMissionByRank(rank);
+        if (missions != null) {
+            return ResponseEntity.ok(missions);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Mission with Rank " + rank + " does not exist in our registration!");
+        }
     }
 
     // Add mission (CREATE)
     @PostMapping("/create")
-    public MissionsDTO createMission(@RequestBody MissionsDTO mission) {
-        return missionsService.createMission(mission);
+    public ResponseEntity<String> createMission(@RequestBody MissionsDTO mission) {
+        MissionsDTO newMission = missionsService.createMission(mission);
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body("Mission created successfully: " + newMission.getName() + ". (ID): " + newMission.getId() + ".");
     }
 
     // Change mission data (UPDATE)
     @PutMapping("/update/{id}")
-    public MissionsDTO updateMission(@RequestBody  MissionsDTO missionUpdated, @PathVariable Long id) {
-        return missionsService.updateMission(missionUpdated, id);
+    public ResponseEntity<?> updateMission(@RequestBody MissionsDTO missionUpdated, @PathVariable Long id) {
+        MissionsDTO mission = missionsService.updateMission(missionUpdated, id);
+        if (mission != null) {
+            return ResponseEntity.ok(mission);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Mission with id: " + id + " does not exist in our registration!");
+        }
     }
 
     // Delete mission (DELETE)
     @DeleteMapping("/delete/{id}")
-    public void deleteMissionById(@PathVariable Long id) {
-        missionsService.deleteMissionById(id);
+    public ResponseEntity<String> deleteMissionById(@PathVariable Long id) {
+        if (missionsService.existsMissionById(id)) {
+            missionsService.deleteMissionById(id);
+            return ResponseEntity.ok("Mission with ID: " + id + " deleted successfully.");
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Mission with ID: " + id + " not found.");
+        }
     }
-
 }
